Split TypeSelectorView into partial files and extract helpers
- Move row binding, search/keyboard input, navigation, the generic-argument
  page flow and view rendering into per-concern partial files; the largest is
  now 302 lines, down from a single 1122-line file
- Extract the footer-hint text into a pure TypeSelectorFooterHint helper
- Share the star/None glyphs through TypeSelectorGlyphs instead of duplicating
  the literals across row rendering and the hint
- No behaviour change; the public API is untouched and Unity compiles clean
