+++ Aspid.FastTools/Packages/tech.aspid.fasttools/Unity/Editor/Scripts/Types/Selectors/TypeSelectorFooterHint.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b762bddc70e8740c7b9ba0c438fb9d5b
+guid: 10204cd029481494f80eafb55972b7c4
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
