+++ Aspid.FastTools/Packages/tech.aspid.fasttools/Unity/Editor/Scripts/Types/Selectors/TypeSelectorGlyphs.cs
@@ -1,0 +1,14 @@
+// ReSharper disable once CheckNamespace
+namespace Aspid.FastTools.Types.Editors
+{
+    /// <summary>
+    /// Unicode glyphs shared between the type selector's row rendering and its footer hint: the favorite
+    /// star (filled/empty) and the &lt;None&gt; marker.
+    /// </summary>
+    internal static class TypeSelectorGlyphs
+    {
+        public const string StarFilled = "★";
+        public const string StarEmpty = "☆";
+        public const string None = "○";
+    }
+}
