+++ Aspid.FastTools/Packages/tech.aspid.fasttools/Unity/Editor/Scripts/Types/Selectors/TypeSelectorGlyphs.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b762bddc70e8740c7b9ba0c438fb9d5b
+guid: 3ad0f0c2264fd4bc783703c622a13fe2
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
