+++ Aspid.FastTools/Packages/tech.aspid.fasttools/Unity/Editor/Scripts/Types/Selectors/TypeSelectorView.Generics.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b762bddc70e8740c7b9ba0c438fb9d5b
+guid: ec81292eedc564a35b1c4fe8b96e0045
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
