+++ Aspid.FastTools/Packages/tech.aspid.fasttools/Unity/Editor/Scripts/Types/Selectors/TypeSelectorView.Input.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b762bddc70e8740c7b9ba0c438fb9d5b
+guid: 09a36821abe4447a786c2c5b67d45d05
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
