+++ Aspid.FastTools/Packages/tech.aspid.fasttools/Unity/Editor/Scripts/Types/Selectors/TypeSelectorView.Navigation.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b762bddc70e8740c7b9ba0c438fb9d5b
+guid: 173b5eda6330a42d8aa9173799cf1b30
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
