+++ Aspid.FastTools/Packages/tech.aspid.fasttools/Unity/Editor/Scripts/Types/Selectors/TypeSelectorView.View.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b762bddc70e8740c7b9ba0c438fb9d5b
+guid: e004097e8bef34fd1988dd8ea7d0fdc8
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
